// apps/admin-web/src/.generated/api/model/maintenanceConfigDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Admin API
 * Control Plane Admin API Service
 * OpenAPI spec version: 1.0.0
 */
import type { RecurringMaintenanceDto } from './recurringMaintenanceDto';
import type { TemporaryMaintenanceDto } from './temporaryMaintenanceDto';

export interface MaintenanceConfigDto {
  /** 임시 점검 설정 */
  temporary: TemporaryMaintenanceDto;
  /** 정기 점검 설정 */
  recurring: RecurringMaintenanceDto;
}